requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
drf-excel==2.2.0
et-xmlfile==1.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
openpyxl==3.0.10
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
unicodecsv==0.14.1
zipp==3.10.0
